make tabs more reusable
