+++ src/SelectMenu/SelectMenuTabs.js
@@ -1,40 +1,14 @@
-import React, {useContext} from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
-import uuid from 'uuid'
-import {tabWrapperStyles, tabStyles} from './SelectMenuStyles'
+import {tabWrapperStyles} from './SelectMenuStyles'
 import {COMMON} from '../constants'
-import {MenuContext} from './SelectMenuContext'
 import theme from '../theme'
 
-const TabBase = ({name, index, ...rest}) => {
-  const menuContext = useContext(MenuContext)
-  const handleClick = () => {
-    menuContext.setSelectedTab(name)
-  }
-  if (!menuContext.selectedTab && index === 0) {
-    menuContext.setSelectedTab(name)
-  }
-
-  const isSelected = menuContext.selectedTab === name
-
+const Tabs = ({className, children, ...rest}) => {
   return (
-    <button role="tab" className="SelectMenuTab" aria-selected={isSelected} onClick={handleClick} {...rest}>
-      {name}
-    </button>
-  )
-}
-
-const SelectMenuTab = styled(TabBase)`
-  ${tabStyles}
-`
-
-const Tabs = ({className, tabs}) => {
-  return (
-    <div role="tablist" className={className}>
-      {tabs.map((tab, index) => (
-        <SelectMenuTab key={uuid()} index={index} name={tab} />
-      ))}
+    <div role="tablist" className={className} {...rest}>
+      {children}
     </div>
   )
 }
